+++ agent/listener/database.py
@@ -2,7 +2,7 @@ import os
 import time
 import sqlite3
 import sys
-import server
+import listener.server
 
 # A module to wrap sqllite3 for use with a small database to store things
 # like checks accross both passive and active sections
@@ -10,7 +10,7 @@ import server
 class DB(object):
 
     def __init__(self):
-        if getattr(sys, u'frozen', False):
+        if getattr(sys, 'frozen', False):
             self.dbfile = os.path.abspath(os.path.dirname(sys.executable) + '/var/ncpa.db')
         else:
             self.dbfile = os.path.abspath(os.path.dirname(__file__) + '/../var/ncpa.db')