+++ agent/passive/nrds.py
@@ -2,12 +2,12 @@ from __future__ import with_statement
 import sys
 import xml.etree.ElementTree as ET
 from passive.nagioshandler import NagiosHandler
-import utils
+import passive.utils
 import tempfile
 import re
 import logging
 import os
-import ConfigParser
+import configparser as cp
 
 
 class Handler(NagiosHandler):
@@ -29,7 +29,7 @@ class Handler(NagiosHandler):
             nrds_config = self.config.get('nrds', 'config_name')
             nrds_config_version = self.config.get('nrds', 'config_version')
             nrds_token = self.config.get('nrds', 'token')
-        except (ConfigParser.NoOptionError, ConfigParser.NoSectionError) as exc:
+        except (cp.NoOptionError, cp.NoSectionError) as exc:
             logging.error("Encountered error while getting NRDS config values: %r", exc)
 
         # Make sure valid input was stated in the config, if not, error out and log it.
@@ -62,7 +62,7 @@ class Handler(NagiosHandler):
         }
 
         # This plugin_abs_path should be absolute, as it is adjusted when the daemon runs.
-        url_request = utils.send_request(nrds_url, **getargs)
+        url_request = passive.utils.send_request(nrds_url, **getargs)
         plugin_abs_path = os.path.join(plugin_path, plugin)
 
         if plugin_abs_path != os.path.abspath(plugin_abs_path):
@@ -73,7 +73,7 @@ class Handler(NagiosHandler):
         try:
             with open(plugin_abs_path, 'wb') as plugin_file:
                 plugin_file.write(url_request)
-                os.chmod(plugin_abs_path, 0775)
+                os.chmod(plugin_abs_path, 0x0775)
         except Exception as exc:
             logging.error('Could not write the plugin to %s: %r', plugin_abs_path, exc)
 
@@ -90,7 +90,7 @@ class Handler(NagiosHandler):
             'token': nrds_token
         }
 
-        nrds_response = utils.send_request(nrds_url, **get_args)
+        nrds_response = passive.utils.send_request(nrds_url, **get_args)
 
         try:
             with tempfile.TemporaryFile() as temp_config:
@@ -98,7 +98,7 @@ class Handler(NagiosHandler):
                 temp_config.write(nrds_response)
                 temp_config.seek(0)
 
-                test_config = ConfigParser.ConfigParser()
+                test_config = cp.ConfigParser()
                 test_config.readfp(temp_config)
 
                 if not test_config.sections():